[WIP] Hybrid engine: use unique table during construction of the hybrid iteration matrix
To prepare for the iterations, a copy of the MTBDD augmented with the relevant ODD information is constructed. Previously, this ensured the uniqueness of the nodes with a linear search among the already constructed nodes for the relevant level. We have observed in several (large) models that this can lead to a significant slow-down.

Here, we use an unordered_map for each variable level to serve as a unique table.

TODO:
 * Evaluate performance/memory impact
 * Properly write output to log instead of using printf